+++ tests/testthat/test-epiweek.R
@@ -304,3 +304,12 @@ test_that("epiweek, miscellaneous work", {
     expect_error(epiweek(year = character()), "`year` must be integer.", fixed = TRUE)
     expect_error(epiweek(week = character()), "`week` must be integer.", fixed = TRUE)
 })
+
+test_that("epiweek boundary functions work", {
+    dates <- as.Date("2020-01-01") + 0:14
+    weeks <- as_epiweek(dates)
+    starts <- as.Date(weeks)
+    ends <- starts + 7 - 1L
+    expect_identical(date_start(weeks), starts)
+    expect_identical(date_end(weeks), ends)
+})