+++ NEWS.md
@@ -1,3 +1,11 @@
+# grates 1.2.0
+
+* New functions `date_start()` and `date_end()` for accessing the
+  boundary elements of `<grates>` objects.
+
+* New function `%during%` for testing whether a scalar date is contained within
+  the range (inclusive) of a `<grates>` object.
+
 # grates 1.1.1
 
 * Users of the 3.5.0 release of